Introduce a common base class for replication and distribution interceptors for code reuse
Move some methods from DistributionManager to BaseDistributionInterceptor.
